requirements.txt
# Core ML/CV Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0  # YOLOv8 for object detection
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Computer Vision & Image Processing
mediapipe>=0.10.0  # For pose estimation
albumentations>=1.3.0  # Data augmentation
segmentation-models-pytorch>=0.3.0  # Segmentation models

# Data Processing
pandas>=2.0.0
scipy>=1.10.0
networkx>=3.0.0  # For graph-based pathfinding
scikit-learn>=1.3.0  # For clustering and preprocessing

# API & Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.4.0

# Database
sqlalchemy>=2.0.0
# sqlite3  # Built-in, no need to install

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Optimization
onnx>=1.15.0
onnxruntime>=1.16.0

# Dataset handling
roboflow>=1.1.0
kaggle>=1.5.0

